BackendCris/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fonttools==4.54.1
idna==3.10
joblib==1.4.2
jsonpickle==3.3.0
kiwisolver==1.4.7
matplotlib==3.9.2
more-itertools==10.5.0
music21==9.1.0
numpy==2.1.2
packaging==24.1
pillow==11.0.0
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
webcolors==24.8.0